6.2 バインド
