+++ src/main/java/com/example/application/service/UserApplicationService.java
@@ -15,13 +15,13 @@ public class UserApplicationService {
 	private MessageSource messageSource;
 	
 	/** 性別のMapを生成する */
-	public Map<String, Integer> getGenderMap() {
+	public Map<String, Integer> getGenderMap(Locale locale) {
 		// キーがString型でバリューがInteger型のMapを生成
 		Map<String, Integer> genderMap = new LinkedHashMap<>();
-		// messagePropertiesのキーmaleのバリューを取得
-		String male = messageSource.getMessage("male", null, Locale.JAPAN);
-		// messagePropertiesのキーfemaleのバリューを取得
-		String female = messageSource.getMessage("female", null, Locale.JAPAN);
+		// MessageSourceクラスのmessageSourceメソッドを使用して、messages.propertiesファイルのmaleというキーのバリューを取得
+		String male = messageSource.getMessage("male", null, locale);
+		// MessageSourceクラスのmessageSourceメソッドを使用して、messages.propertiesファイルのfemaleというキーのバリューを取得
+		String female = messageSource.getMessage("female", null, locale);
 		// 変数genderMapにキーがmale、バリューが1のデータを追加
 		genderMap.put(male, 1);
 		// 変数genderMapにキーがfemale、バリューが2のデータを追加
