+++ src/main/java/com/example/controller/LoginController.java
@@ -1,4 +1,4 @@
-package com.example.application.controller;
+package com.example.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
