+++ src/main/resources/application.properties
@@ -16,3 +16,6 @@ spring.datasource.data=classpath:data.sql
 
 #H2DBのコンソール画面をブラウザで表示するための設定
 spring.h2.console.enabled=true
+
+# message
+spring.messages.basename=i18n/messages
