+++ src/main/resources/i18n/messages_en.properties
@@ -1,0 +1,10 @@
+user.signup.title=Signup
+user.signup.btn=Signup
+userId=User ID
+password=Password
+userName=User Name
+birthday=Birthday
+age=Age
+male=Male
+female=Female
+gender=Gender
